BF: Use cassette should skip_if_no_network
move skip_if_no_network to vcr to not have an import cycle
broken:
  skip_if_no_network can't wrap around vcr.use_cassette as it returns a
  decorator object instead of a function

Issue #700
